# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vehicle Management API",
        "description": "API documentation for Vehicle Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        },
        {
            "url": "http://localhost/api",
            "description": "Docker API Server"
        }
    ],
    "paths": {
...
        "/vehicle-attributes": {
            "get": {
                "tags": [
                    "Vehicle Attributes"
                ],
                "summary": "Get all vehicle attributes",
                "operationId": "5a70a309249e0159124b38e324c91de4",
                "responses": {
                    "200": {
                        "description": "List of vehicle attributes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "top_speed"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Maximum speed of the vehicle"
                                            },
                                            "data_type": {
                                                "type": "string",
                                                "example": "float"
                                            },
                                            "unit": {
                                                "type": "string",
                                                "example": "km/h"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vehicle Attributes"
                ],
                "summary": "Create a new vehicle attribute",
                "operationId": "9764f7739851bf7a686093da61f2db43",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "data_type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "battery_capacity"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Battery capacity for electric vehicles"
                                    },
                                    "data_type": {
                                        "type": "string",
                                        "example": "float"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "kWh"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Attribute created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vehicle-attributes/{id}": {
            "get": {
                "tags": [
                    "Vehicle Attributes"
                ],
                "summary": "Get a specific vehicle attribute",
                "operationId": "f24de8c2a3ae76e5f8161e3c06159585",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle attribute details"
                    },
                    "404": {
                        "description": "Attribute not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Vehicle Attributes"
                ],
                "summary": "Update a vehicle attribute",
                "operationId": "0ce85fe765403b44e9ee3b1bc41403c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "data_type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "battery_capacity"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "data_type": {
                                        "type": "string",
                                        "example": "float"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "kWh"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Attribute updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Vehicle Attributes"
                ],
                "summary": "Delete a vehicle attribute",
                "operationId": "73583723d1bb082ee6c5170dd1ee967d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Attribute deleted successfully"
                    },
                    "404": {
                        "description": "Attribute not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}